requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
PyYAML
websockets
pytest
httpx
pytest-asyncio
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
jupyter
jupyterlab
nbformat
ipywidgets
openpyxl
xlrd
lxml
beautifulsoup4
html5lib
requests
tqdm
loguru
watchdog
schedule
APScheduler
celery
redis
flower
kafka-python
confluent-kafka
sqlalchemy
alembic
psycopg2-binary
pymysql
sqlite-utils
django
django-rest-framework
flask
flask-restx
fastapi-users
passlib[bcrypt]
python-jose[cryptography]
paho-mqtt
aiohttp
aiohttp-retry
gymnasium